+++ include/crc.h
@@ -87,11 +87,11 @@ static inline uint32_t crc32_final(uint32_t crc)
 }
 
 bool crc32_validate(uint32_t expected_crc, const uint8_t header[ENTRY_HEADER_SIZE],
-                    const uint8_t *key, size_t key_size,
-                    int fd, off_t value_pos, size_t value_size);
+                    const uint8_t *key, uint32_t key_size,
+                    int fd, uint32_t value_pos, uint32_t value_size);
 
 bool crc32_validate_buf(uint32_t expected_crc, const uint8_t header[ENTRY_HEADER_SIZE],
-                        const uint8_t *key, size_t key_size,
-                        const uint8_t *value, size_t value_size);
+                        const uint8_t *key, uint32_t key_size,
+                        const uint8_t *value, uint32_t value_size);
 
 #endif
